Rename tour steps for better reporting
