+++ packages/haiku-creator/src/react/components/Tour/Steps/Congratulations.js
