+++ packages/haiku-creator/src/react/components/Tour/Steps/index.js
@@ -1,5 +1,6 @@
 export { default as AddState } from './AddState'
-export { default as AnimatorNotice } from './AnimatorNotice'
+export { default as Congratulations } from './Congratulations'
+export { default as CreateTween } from './CreateTween'
 export { default as Finish } from './Finish'
 export { default as LibraryStart } from './LibraryStart'
 export { default as ModifyProperty } from './ModifyProperty'
@@ -12,5 +13,4 @@ export { default as PublishedLink } from './PublishedLink'
 export { default as ReferenceState } from './ReferenceState'
 export { default as StatesStart } from './StatesStart'
 export { default as Summonables } from './Summonables'
-export { default as TweenCreator } from './TweenCreator'
 export { default as Welcome } from './Welcome'
